build flag to enable equality deletes is not only in a few places
