+++ src/core/metadata/manifest/iceberg_manifest.cpp
@@ -308,14 +308,12 @@ Value IcebergDataFile::ToValue(const IcebergTableMetadata &table_metadata, const
 	}
 	children.push_back(Value::MAP(LogicalType::STRUCT(null_value_count_types), null_value_counts_values));
 
-#ifdef ICEBERG_ENABLE_EQUALITY_DELETE_WRITES
 	// equality_ids: list<int> (empty for non equality-delete files)
 	vector<Value> equality_id_values;
 	for (auto &id : equality_ids) {
 		equality_id_values.push_back(Value::INTEGER(id));
 	}
 	children.push_back(Value::LIST(LogicalType::INTEGER, std::move(equality_id_values)));
-#endif
 
 	// referenced_data_file
 	if (table_metadata.iceberg_version >= 3) {
@@ -734,7 +732,6 @@ idx_t WriteToFile(const IcebergTableMetadata &table_metadata, const IcebergManif
 		yyjson_mut_obj_add_strcpy(doc, val_obj, "type", "binary");
 		yyjson_mut_obj_add_uint(doc, val_obj, "id", NULL_VALUE_COUNTS_VALUE);
 	}
-#ifdef ICEBERG_ENABLE_EQUALITY_DELETE_WRITES
 	// equality_ids: list<int> - the field-ids an equality-delete file applies to
 	{
 		children.emplace_back("equality_ids", LogicalType::LIST(LogicalType::INTEGER));
@@ -755,7 +752,6 @@ idx_t WriteToFile(const IcebergTableMetadata &table_metadata, const IcebergManif
 		yyjson_mut_obj_add_strcpy(doc, list_type, "element", "int");
 		yyjson_mut_obj_add_bool(doc, list_type, "element-required", true);
 	}
-#endif
 	// referenced_data_file
 	if (table_metadata.iceberg_version >= 3) {
 		// referenced_data_file: long
